First steps at parsing the old listing.
The new class listing2xml tries to extract comments and (some)
symbols from the disassembled ROM listing in the resources.

Unfortunately there are no symbols at all in the listing.
There have been <= 8 character names for every ROM function
been known for the TRS80, which would in part be usable for
the Coulour Genie, because the first 12K are more or less
a hacked copy of the TRS80 ROM.