WIP listing2xml parser almost done; TRS80 glyphs.
Finally the listing2xml parser is producing a halfway usable
definitions file (cgenie-new.xml).

Add an abstraction for bdfData and derive bdfCgenie and bdfTrs80 from it.
The TRS80 BDF file is not yet usable. Need to fix the calculations of
bounding boxes, capheight, xheight, etc. and finally convert the
result to TTF using FontForge.